rename & re-organize files to better reflect the underlying architecture
