+++ test/unit/Cardano/Wallet/Network/HttpBridgeSpec.hs
