+++ test/unit/Cardano/Wallet/Primitive/MnemonicSpec.hs
